Refactor chartPatterns.ts to separate data and logic
